+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 * Added `sheet-name?` predicate.
 * Added `auto-size-column!` and `auto-size-all-columns!` functions to automatically adjust column widths to contents.
 * Added `column-index-seq` function to get the indices of the columns in a row.
-* Expose `:quote-prefixed` cell style option in `create-cell-style!`
+* Exposed `:quote-prefixed` cell style option in `create-cell-style!`
 
 ## Version 1.18.0